Avoid remove remote error
Fixes APP-16W

+++ lib/setup/screens.dart
@@ -440,7 +440,9 @@ class GitHostSetupScreenState extends State<GitHostSetupScreen> {
         await GitRepository.init(repoPath);
       }
       var repo = await GitRepository.load(repoPath).getOrThrow();
-      await repo.removeRemote(widget.remoteName).throwOnError();
+      if (repo.config.remote(widget.remoteName) != null) {
+        await repo.removeRemote(widget.remoteName).throwOnError();
+      }
     } on Exception catch (e, stacktrace) {
       Log.e("Failed to remove remote", ex: e, stacktrace: stacktrace);
       logExceptionWarning(e, stacktrace);
